backend/requirements.txt
acres==0.3.0
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
argcomplete==1.10.3
bcrypt==4.2.1
beautifulsoup4==4.8.2
cachetools==5.5.1
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
compressed-rtf==1.0.6
configobj==5.0.9
configparser==7.1.0
DateTime==5.5
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
ebcdic==1.1.1
email_validator==2.2.0
etelemetry==0.3.1
exceptiongroup==1.2.2
extract-msg==0.28.7
fastapi==0.115.7
fastapi-cli==0.0.7
filelock==3.17.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.5
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
olefile==0.47
openai==1.60.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.10.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scipy==1.15.2
shellingham==1.5.4
simplejson==3.20.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
starlette==0.45.3
textract==1.6.5
tiktoken==0.8.0
tqdm==4.67.1
traits==7.0.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
xlrd==1.2.0
XlsxWriter==3.2.2
zope.interface==7.2
